+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/model/response/PolicyResponse.java
@@ -17,7 +17,21 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-export type NotificationTabInformation = {
-  tabIndex: string | null,
-  pageNumber?: number | null
+package org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response;
+
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.model.IrsPolicy;
+
+import java.time.Instant;
+
+public record PolicyResponse(
+        String policyId,
+        Instant createdOn,
+        Instant validUntil
+) {
+    public IrsPolicy toDomain() {
+        return IrsPolicy.builder()
+                .policyId(this.policyId)
+                .ttl(this.validUntil.toString())
+                .build();
+    }
 }